Create a module to work with markdown
